In visitSTORE, always use FindBetterChain, rather than only when UseAA is enabled.
Also, simplify the MergeConsecutiveStores to only use the mode where
it looks for stores attached to the same chain node (previously also
restricted to UseAA), rather than looking for a sequence of stores
chained each to the next.

This finishes the change Matt Arsenault started in r246307.

The main reason for this change is that the new code is conceptually
cleaner: it separates out the decisions about which stores and loads
are non-interfering, and can have their chain dependency relaxed, from
the decision about merging non-interfering stores. The way the old
code detected non-overlapping stores was sketchy, and subtly broken
multiple times.

Further, the new way usually manages to produce better code. This, for
two reasons:

1. It enables loads to be forwarded from the stored value more often,
eliminating the loads entirely, because the chain predecessor of the
load is more often the related store, with no intervening operations.

2. It enables merging to occur more reliably, because the chain was
already sometimes organized differently than a linear dependence.

One situation exists now where the new code is not able to detect
merge candidates and it was before: when some stores overlap a load,
and others do not. This causes
test/CodeGen/X86/merge-store-partially-alias-loads.ll to no longer
work.

A couple other minor related changes were made at the same time:

1) Re-add the Store node to the worklist after calling
   SimplifyDemandedBits.

2) Allow value forwarding from a Store, through a TokenFactor, to the
   eventual Load, rather than only directly from Store->Load. This is
   more important now than it was before, since the chain
   rearrangement can introduce TokenFactor nodes.

3) Increase GatherAllAliasesMaxDepth from 6 to 18. That number is
   arbitrary, but seemed sufficient to not cause regressions in tests.

I had to adjust a number of test cases to take account of the
arbitrarily different ordering. A few of the tests deserve explicit note:
- CodeGen/AArch64/argument-blocks.ll:
   It's not entirely clear what the test_varargs_stackalign test is
   supposed to be asserting, but I think the new code looks right.

- CodeGen/AArch64/arm64-alloc-no-stack-realign.ll:
   The emitted code on this got worse (str+stp+str, instead of
   stp+stp). I believe this is simply due to the target-specific
   AArch64LoadStoreOptimizer pass being sensitive to the arbitrary
   order of the four unaliased stores, and thus merges the middle two
   instead of the first two and last two.

- CodeGen/AArch64/arm64-dagcombiner-dead-indexed-load.ll
   This test had failed without the change to re-add the current node
   to the worklist after SimplifyDemandedBits.

- CodeGen/AArch64/arm64-memset-inline.ll
  CodeGen/AArch64/arm64-stur.ll
  CodeGen/ARM/memset-inline.ll
   The backend now generates *worse* code due to store merging
   succeeding, as it doesn't do a 16-byte constant-zero store
   efficiently.

- CodeGen/AArch64/merge-store.ll
   Improved, but there still seems to be an extraneous vector insert
   from an element to itself?

- CodeGen/AMDGPU/merge-stores.ll
   Has improved, but, the now-properly-merged v8i32 constant store is
   re-expanded into 32-bit stores, instead of doing a single 128bit
   store.

- CodeGen/PowerPC/ppc64-align-long-double.ll
   Worse code emitted in this case, due to the improved store->load
   forwarding.

- CodeGen/X86/dag-merge-fast-accesses.ll:
- CodeGen/X86/MergeConsecutiveStores.ll:
- CodeGen/X86/stores-merging.ll:
   Restored correct merging of non-aligned stores.

- CodeGen/X86/merge-store-partially-alias-loads.ll:
   Partially overlapping stores no longer merged. :(

- CodeGen/X86/pr18023.ll:
   This test has been removed, as it was asserting incorrect
   behavior. Non-volatile stores *CAN* be moved past volatile loads,
   and now are.