Support Intel i386 and i486 CPUs.
As far as normal instructions go, only a few need to be excluded: the
the 486 doesn't have CMPXCHG8B, and 386 is additionally missing BSWAP,
CMPXCHG, and XADD.

Previously, llvm would emit these instructions even if you asked to
build for a 386.

XADD and CMPXCHG are used only for atomics; we can now ask
AtomicExpandPass to deal with expanding to libcalls on these CPUs, so it
is trivial to avoid emitting those.

BSWAP can be implemented via 3 RORs, instead. (This custom expansion is
much better than the default you get if you use "Expand")